Ditch the 'fetch' option and rework a8bfd9b2efbba9aaa57049efd45a3e0098f1d917
Being able to dynamically switch the parser with sub-ref per-request
seems very clever, thus is probably unwanted.

For using Hijk with Hijk::HTTP::XS, we do:

    use Hijk;
    use Hijk::HTTP::XS;

Or in some dynamic fashion:

    use Hijk;
    require Hijk::HTTP::XS if MY_GUT_SAYS_SO;

While a sole 'use Hijk;' should always means to use only the pure-perl version.

This keeps it simple for developer to make decisions and there should
be no surprises on what versions are used. (Well, unless there are some code
act in distance, but that applies generically.)