Remove the Hijk::HTTP::XS module
We've found that the gains in doing this XS stuff really aren't
noticeable at all on production workloads, since most of the time is I/O
bound. This resolves issue #4 by just getting rid of the XS module,
talked to Borislav about it and he suggested just removing it, so doing
that.