Fix a bad bug in the chunked encoding and disable it by default
I don't think we should be enabling this by default while there's still
basic issues like what I'm fxing here being found in the code (see the
changelog).

In any case even if this was proven to be bug-free I think disabling it
is the right move, Hijk is primarily going to be used for internal
services where you control both ends, for small responses the overall
best solution is usually just to disable chunked encoding on the other
end, and save the client from doing the work to do the extra parsing.