Fix what turned out to be a major regression reported in RT #101424
We didn't spot this because we had no live tests that made enough
requests to kick in "Connection: close" on the remote end, now we have
that.

Note how early in the do {} block I've added a return of
$connection_close, I did this just for completeness, but I'm not aware
of any case where that would actually cause us to return an true
$connection_close.

Unifying some of this logic I added in 0.12-45-g972c2f2 also makes for
simpler code that doesn't branch as much depending on $head_as_array.