Amend the custom Host feature and document + test it
As discussed in pull request #22 I think this is better than the patch
submitted there, because:

 * You could completely omit the "Host" header this way, which is vali
   HTTP/1.0.

 * Even though HTTP/1.1 mandates it Hijk is a "raw" request library, you
   should be able to pass custom requests that don't have it.

 * We don't need to loop through the headers looking for "host" headers
   every time we construct a request.

 * This also allows you to control where in the header list the Host
   header is, again for the "raw" request part.

The downside is that it doesn't Just Work and you need to pass the
send_host_header => 0 option, but I think that's fine. This is an
obscure enough feature that finding & passing that one flag should be
fine, and we still DWYM by passing "Host" by default.