live-plack.t: disable TEST_LIVE==1 guard.
This test does not depend on any externally-controlled resources and
can be run entierly at localhost. It also setup/teardown its own
resources (plack app), meaning that there is no uncertainty about
whether an external service is alive or dead. For this reason, there
is no need to guard it with this variable.

also rename the test for the same reason.