depend on HTTP::Server::Simple::PSGI server on its control Content-Length header.
Well, not reallly control.

Exsting PSGI servers behave differently when it comes to controlling
the value of Content-Length header, most implementation calculates the
value should the message body is determined, and none of them passthru
the values given by the inner application.

It also seems to be the case that all existing PSGI server will change
the Content-Length header to "0", if the request method is "HEAD" and
the inner-app sets a non-zero value to Content-Length.

Qouting from this old RFC, it looks like the value could be non-zero:

    https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.13

    The Content-Length entity-header field indicates the size of the
    entity-body, in decimal number of OCTETs, sent to the recipient
    or, in the case of the HEAD method, the size of the entity-body
    that would have been sent had the request been a GET.

However, if it is never the case de facto, there is probably no point
to arg with a phanthom spec.

Until next time, when we meet a HTTP server in the wild...